.long-string error LINE # points to directive
Source#pop_line always changes the current_line pointer.  This caused a
missing .end-long-string error to reference the last line in the file
because the .long-string directive would pop all the lines in the file.
To fix this I added Source#pop_sub_line that does not change the
current_line.